# docs/swagger/swagger.yaml
basePath: /
definitions:
  APIErrorInfo:
    properties:
      code:
        example: validation_error
        type: string
      details:
        additionalProperties: true
        type: object
      message:
        example: Validation failed
        type: string
    type: object
  APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/APIErrorInfo'
      success:
        example: true
        type: boolean
      timestamp:
        example: "2025-01-15T10:30:00Z"
        type: string
    type: object
...
  CreateWebhookRequest:
    properties:
      events:
        description: Se vazio, aceita todos os eventos
        items:
          type: string
        type: array
      secret:
        type: string
      url:
        type: string
    required:
    - url
    type: object
...
  ErrorResponse:
    properties:
      error:
        example: validation_error
        type: string
      message:
        example: name is required
        type: string
    type: object
  EventCategoryResponse:
    properties:
      category:
        type: string
      events:
        items:
          type: string
        type: array
    type: object
...
  ListEventsResponse:
    properties:
      allEvents:
        items:
          type: string
        type: array
      categories:
        items:
          $ref: '#/definitions/EventCategoryResponse'
        type: array
    type: object
...
  WebhookResponse:
    properties:
      createdAt:
        type: string
      enabled:
        type: boolean
      events:
        items:
          type: string
        type: array
      id:
        type: string
      sessionId:
        type: string
      updatedAt:
        type: string
      url:
        type: string
    type: object
...
info:
  contact:
    email: support@zpwoot.com
    name: zpwoot Support
    url: https://github.com/your-org/zpwoot
  description: |-
    A comprehensive WhatsApp Business API built with Go, following Clean Architecture principles.
    Provides endpoints for session management, messaging, contacts, groups, media handling, and integrations.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: zpwoot WhatsApp API
  version: 1.0.0
paths:
...
  /sessions/{sessionId}/webhooks:
    delete:
      description: Delete webhook configuration for a session
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Webhook deleted successfully
          schema:
            $ref: '#/definitions/APIResponse'
        "404":
          description: Webhook not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Webhook Configuration
      tags:
      - Webhooks
    get:
      description: Get webhook configuration for a session
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Webhook configuration
          schema:
            $ref: '#/definitions/WebhookResponse'
        "404":
          description: Webhook not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Webhook Configuration
      tags:
      - Webhooks
    post:
      consumes:
      - application/json
      description: Configure or update webhook for a session
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Webhook configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreateWebhookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Webhook configured successfully
          schema:
            $ref: '#/definitions/WebhookResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Configure Webhook
      tags:
      - Webhooks
  /webhooks/events:
    get:
      description: List all available webhook event types
      produces:
      - application/json
      responses:
        "200":
          description: Available events
          schema:
            $ref: '#/definitions/ListEventsResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List Available Events
      tags:
      - Webhooks
securityDefinitions:
  ApiKeyAuth:
    description: 'API key for authentication. Use the value from .env file (API_KEY
      variable). Send as ''Authorization: your-api-key'' (without Bearer prefix)'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...